requirements.txt
crewai
speechrecognition
coqui-tts
pydantic